+++ src/common/components/Gw2Map/index.js
@@ -1,4 +1,6 @@
 import { PropTypes } from 'react';
+import cx from 'classnames';
+
 import styles from './styles.less';
 
 function getStyle (id = 0) {
@@ -9,14 +11,15 @@ function getStyle (id = 0) {
   };
 }
 
-const Gw2Map = ({ data }) => (
-  <div className={styles.root} style={getStyle(data.id)}>
+const Gw2Map = ({ data, className }) => (
+  <div className={cx(styles.root, className)} style={getStyle(data.id)}>
     {data.name && <span title={data.name} className={styles.name}>{data.name}</span>}
   </div>
 );
 
 Gw2Map.propTypes = {
   data: PropTypes.object,
+  className: PropTypes.string,
 };
 
 Gw2Map.defaultProps = {